backend/requirements.txt
google-genai==1.11.0
python-dotenv==1.1.0
fastapi==0.115.12
pydantic==2.11.3
uvicorn==0.34.2
weasyprint==65.1